requirements.txt
openai
langchain
langchain-openai
langchain-community
chromadb
streamlit
pypdf
python-dotenv
tiktoken